feat: use normal properties to avoid function calls to proxy objects
