+++ blockscipy/src/chain/tx/tx_py.cpp
@@ -7,20 +7,22 @@
 //
 
 #include "tx_py.hpp"
+#include "tx_properties_py.hpp"
 #include "ranges_py.hpp"
 #include "caster_py.hpp"
+#include "self_apply_py.hpp"
 
 #include <blocksci/chain/access.hpp>
 #include <blocksci/chain/blockchain.hpp>
-#include <blocksci/cluster/cluster.hpp>
-
-#include <pybind11/operators.h>
+#include <blocksci/chain/block.hpp>
 
 namespace py = pybind11;
 
 using namespace blocksci;
 
 void init_tx(py::class_<Transaction> &cl) {
+    applyMethodsToSelf(cl, AddTransactionMethods{});
+
     cl
     .def("__str__", &Transaction::toString)
     .def("__repr__", &Transaction::toString)
