+++ test/blockscipy/test_chain.py
@@ -87,7 +87,20 @@ def test_filter_blocks(chain):
     assert blocks1 == blocks2
 
 
+def test_filter_blocks_legacy(chain):
+    blocks1 = [b for b in chain.blocks if b.height % 3 == 0]
+    blocks2 = chain.filter_blocks_legacy(lambda b: b.height % 3 == 0)
+
+    assert blocks1 == blocks2
+
+
 def test_filter_txes(chain):
     txs1 = [tx for block in chain for tx in block if tx.block.height % 3 == 0]
     txs2 = chain.filter_txes(lambda tx: tx.block.height % 3 == 0)
+    assert txs1 == txs2
+
+
+def test_filter_txes_legacy(chain):
+    txs1 = [tx for block in chain for tx in block if tx.block.height % 3 == 0]
+    txs2 = chain.filter_txes_legacy(lambda tx: tx.block.height % 3 == 0)
     assert txs1 == txs2
